Added a new flag to main.go
This flag is used to set the day that shall be processed.
Also renamed the download flag to "-g" for "get" to avoid command shadowing.

If a day is set, runDay will be called with the provided number, otherwise we'll loop through all the days running each of them in order.